+++ packages/lens-react/lib/components/Datatable/components/Datatable.tsx
@@ -1,33 +1,45 @@
 import React from "react";
 import { DataTableLensContextProvider } from "../context/DataTableContext";
 import DatatableList from "./DatatableList";
-import { TableStylingInterface } from "../interface/TableStylingInterface";
+import type { TableStylingInterface } from "../interface/TableStylingInterface";
+import type { ActionsMappingPropInterface } from "../interface/ActionsMappingInterface";
+import type { DatatableVariants } from "../context/ContextInterfaces";
+import "../styles.css";
 
 export interface DatatableInterface {
-  onSelect: (selected: any[]) => void;
-  selectedItems: string[];
+  onSelect?: (selected: any[]) => void;
+  selectedItems?: string[];
   classNames?: TableStylingInterface;
   disableResizing?: boolean;
+  disableReordering?: boolean;
   viewId?: string;
-  rowActions?: any;
+  cellActions?: any;
+  actionsMapping?: ActionsMappingPropInterface;
+  variant?: DatatableVariants;
 }
 
 const Datatable: React.FC<DatatableInterface> = ({
-  selectedItems,
-  onSelect,
+  selectedItems = [],
+  onSelect = () => {},
   classNames,
   disableResizing = false,
+  disableReordering = false,
   viewId,
-  rowActions,
+  cellActions,
+  actionsMapping,
+  variant,
 }) => {
   return (
     <DataTableLensContextProvider
       selectedItems={selectedItems}
       onSelect={onSelect}
       classNames={classNames}
-      disableResizing={disableResizing}
       viewId={viewId}
-      rowActions={rowActions}
+      cellActions={cellActions}
+      actionsMapping={actionsMapping}
+      variant={variant}
+      disableResizing={disableResizing}
+      disableReordering={disableReordering}
     >
       <DatatableList />
     </DataTableLensContextProvider>