+++ packages/lens-react/lib/components/Datatable/components/DatatableHeaderItem.tsx
@@ -20,10 +20,10 @@ const DatatableHeaderItem = ({
 }: any) => {
   const fixed = (header.column.columnDef.meta as any)?.fixed;
   const { column } = header;
+  const { classNames, disableResizing, variantClasses, disableReordering } =
+    useDatatableContext();
 
-  const { classNames, disableResizing } = useDatatableContext();
-
-  const css = getColumnPinningStyles(column, true);
+  const css = getColumnPinningStyles(column);
 
   const { attributes, isDragging, listeners, setNodeRef, transform } =
     useSortable({
@@ -44,9 +44,10 @@ const DatatableHeaderItem = ({
   const { minSize, maxSize } = header.column.columnDef;
   const enableResizeHandler =
     minSize !== undefined && maxSize !== undefined ? minSize !== maxSize : true;
-
+  const isLast = column.getIsLastColumn();
   const align = (header.column.columnDef.meta as any)?.align;
   const styles = getStyleClasses(align);
+  const isAction = column.id === "actions";
 
   return (
     <SortableContext
@@ -57,9 +58,14 @@ const DatatableHeaderItem = ({
       <div
         key={header.id}
         className={cn(
-          "relative truncate px-2 py-2 cursor-grab backdrop-blur-[200px]",
+          "relative truncate px-2 py-2",
+          variantClasses.header_cell,
+          !disableReordering && enableResizeHandler
+            ? "cursor-grab"
+            : "cursor-pointer",
           styles.text,
-          classNames && classNames.header
+          classNames && classNames.header,
+          (isAction && classNames?.actionsHeader) || ""
         )}
         style={{
           width: `calc(var(--header-${id}-size) * 1px)`,
@@ -69,6 +75,7 @@ const DatatableHeaderItem = ({
         ref={setNodeRef}
         {...attributes}
         {...listeners}
+        data-islast={isLast}
       >
         {header.isPlaceholder
           ? null
@@ -79,6 +86,7 @@ const DatatableHeaderItem = ({
             isResizing={header.column.getIsResizing()}
             setIsInResizeArea={setIsInResizeArea}
             disabled={!enableResizeHandler}
+            isLast={isLast}
           />
         )}
       </div>