+++ packages/lens-react/lib/components/Datatable/context/DataTableContext.tsx
@@ -12,6 +12,7 @@ import { useViewContext } from "@/components/Views/View";
 import { initialiseDefaultDatatableValues } from "../utils/initialiseDefaultDatatableValues";
 import { initialiseDefaultColumnsConfig } from "../utils/initialiseDefaultColumnsConfig";
 import { initialiseDatatableStates } from "../utils/initialiseDatatableStates";
+import { fetchStylesFromVariants } from "../hooks/fetchStylesFromVariants";
 
 const DatatableContext = createContext<DatatableContextType | undefined>(
   undefined
@@ -23,6 +24,7 @@ const DatatableContextProvider: React.FC<DatatableContextProviderInterface> = ({
   columns,
   selectedItems,
   viewChildRef,
+  variant = "vanilla",
   ...props
 }) => {
   const { defaultColPinning, defaultColShow, tableSelectedItems } =
@@ -100,13 +102,20 @@ const DataTableLensContextProvider: React.FC<
   classNames,
   disableResizing = false,
   viewId = "default",
-  rowActions,
+  cellActions,
+  actionsMapping,
+  variant,
+  disableReordering,
 }) => {
   const lensContext = useLensContext();
   const viewContext = useViewContext();
   const sensors = initialiseDnDSensors();
+  const DATA_TABLE_STYLING_CLASSES = fetchStylesFromVariants(variant);
 
-  const { isFetched, isError, endpoints, modal_name } = lensContext;
+  const { isFetched, isError, endpoints, modal_name, lensConfiguration } =
+    lensContext;
+  const { dataCallback, permissionHeaders } = lensConfiguration;
+
   const {
     filters,
     searchQuery,
@@ -116,10 +125,10 @@ const DataTableLensContextProvider: React.FC<
   } = viewContext;
   const viewName = config?.name || "default_view";
 
-  if (rowActions) {
+  if (cellActions) {
     const visibleAttributes =
       config?.columns.map((column: any) => column?.id) || [];
-    Object.keys(rowActions).forEach((key) => {
+    Object.keys(cellActions).forEach((key) => {
       if (!visibleAttributes.includes(key)) {
         console.error(
           `Table does not contain the attribute: "${key}"`,
@@ -142,7 +151,12 @@ const DataTableLensContextProvider: React.FC<
     columns,
     identifierKey = "",
     allowedScopes,
-  } = useTableConfig(tableConfig);
+  } = useTableConfig(
+    tableConfig,
+    actionsMapping,
+    DATA_TABLE_STYLING_CLASSES,
+    permissionHeaders
+  );
 
   return (
     <DatatableContextProvider
@@ -154,17 +168,21 @@ const DataTableLensContextProvider: React.FC<
       onSelect={onSelect ? onSelect : () => {}}
       selectedItems={selectedItems || []}
       classNames={classNames}
-      disableResizing={disableResizing}
       filters={filters}
       searchQuery={searchQuery}
       viewId={viewId}
       modalName={modal_name}
       viewChildRef={viewChildRef}
-      rowActions={rowActions}
+      cellActions={cellActions}
       viewName={viewName}
       expand={expand}
       localContext={localContext}
       allowedScopes={allowedScopes}
+      variant={variant}
+      variantClasses={DATA_TABLE_STYLING_CLASSES}
+      dataCallback={dataCallback}
+      disableResizing={disableResizing}
+      disableReordering={disableReordering}
     >
       {isFetched ? (
         isError ? (
@@ -185,7 +203,7 @@ const DatatableLoader = () => {
     <div className="w-full h-full  flex flex-col items-center justify-center gap-y-2">
       <div className="relative flex items-center justify-center w-20 h-20 border-4 border-white rounded-full opacity-100 border-t-gray-600 animate-spin"></div>
       <label className="text-lg font-semibold">
-        Lens Powered Dtatatable is rendering your configurations. Please be
+        Lens Powered Datatable is rendering your configurations. Please be
         patient..
       </label>
     </div>